- app: LinkGenerator overhaul, less general, more concrete
- app: introduced Cursor for SequencingState

- channels: upgrade to support new LinkGenerator and Cursor

+++ iota-streams-app/src/message/hdf.rs
@@ -74,8 +74,8 @@ impl<Link> HDF<Link> {
         self.payload_frame_count
     }
 
-    pub fn with_seq_num(mut self, seq_num: u64) -> Self {
-        self.seq_num = Uint64(seq_num);
+    pub fn with_seq_num(mut self, seq_num: u32) -> Self {
+        self.seq_num = Uint64(seq_num as u64);
         self
     }
